move MD flow to separate module
